pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "datathon"
version = "0.1.0"
description = "Datathon FIAP "
authors = [{ name = "Leonardo Pereira", email = "lleonardo_pereira@outlook.com" }]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pandas",
    "mlflow",
    "lightgbm",
    "imbalanced-learn",
    "scikit-learn",
    "SQLAlchemy",
    "python-dotenv",
    "psycopg2-binary",
    "flask"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["datathon_package*"]
